Add num location asserts

+++ programs/helium-entity-manager/src/state.rs
@@ -57,5 +57,5 @@ pub struct IotHotspotInfoV0 {
   pub elevation: Option<i32>,
   pub gain: Option<i32>,
   pub is_full_hotspot: bool,
-  pub num_location_asserts: u32,
+  pub num_location_asserts: u16,
 }
